+++ lua/catppuccin/groups/integrations/ufo.lua
@@ -1,0 +1,10 @@
+local M = {}
+
+M.get = function()
+	return {
+		UfoFoldedFg = { fg = C.lavender },
+		UfoFoldedEllipsis = { fg = C.crust, bg = C.blue },
+	}
+end
+
+return M